✨ [김채영] double형으로 휴가 개수 지급 & 메인에 휴가 지급 출력

+++ src/main/java/com/fiveLink/linkOffice/vacation/service/VacationService.java
@@ -128,7 +128,7 @@ public class VacationService {
 
 
     //1년미만 재직자 한달에 한번 휴가 지급
-    public void incrementVacation(Long memberNo, int num){
+    public void incrementVacation(Long memberNo, double num){
             Member member = memberRepository.findById(memberNo)
                     .orElseThrow(() -> new IllegalArgumentException("직원을 찾을 수 없습니다"));
             // 휴가 일수 증가
@@ -143,7 +143,7 @@ public class VacationService {
     }
 
     //1년이상 재직자 입사일 기준 경과 시, 휴가 지급
-    public void resetVacation(Long memberNo, int num){
+    public void resetVacation(Long memberNo, double num){
 
             Member member = memberRepository.findById(memberNo)
                     .orElseThrow(() -> new IllegalArgumentException("직원을 찾을 수 없습니다"));
@@ -184,7 +184,7 @@ public class VacationService {
     }
 
     //사용자 별 휴가 개수
-    public int userVacationCount(Long memberNo) {
+    public double userVacationCount(Long memberNo) {
         return vacationMapper.userVacationCount(memberNo);
     }
